requirements.txt
pyspark==4.0.1
boto3==1.40.59
botocore==1.40.59
python-dotenv>=1.0.0